Added persistent data storage
Added two functions: one which save the info of the player postion, coins and markers, which and the other which loads the info to set all the data as it is in saved local storage.

Utilized the saving of the game state at the end of every action, like moving, retrieving and deposting coins so that it is constantly saving information.

Added a check which checks if there is anything in local storage, which if there is, it will load that information thus restoring the previous player position, coins and markers