Added Reset Button
Added the reset button, which uses the reset function, which erases the game information in local storage. It also disables geotracking if it was enabled. Erases all Player geo coins and resets player postion. Also added a prompt for the user for double comfirmation regarding resetting

NOTE: While the playercoins are still saved, they are currently not displayed unless the player retrieves or desposits a coin, in which case the player coins is displayed again