D3.D Complete
Added an array that tracks player movement,  which is updated every time the player moves. Used a polyline to render the player moves, by connecting the dots from previous to current location. Lastly, added the player movement history and polyline to the save game state in local storage to preserve the rendered line when loading